[5장_이지훈] DiscountCondtion 분리
